Move page schema into new file
